refactor: fake api data sort function to new file
